tests/e2e/pyproject.toml
[project]
name = "estategap-e2e"
version = "0.1.0"
description = "External end-to-end test suite for EstateGap"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "mypy>=1.11",
  "nats-py>=2.6",
  "pydantic>=2.8",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-xdist>=3.6",
  "redis>=5.0",
  "ruff>=0.6",
  "websockets>=12.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
